✨ Refactor imports in solver job modules and add new model converters
