+++ services/api-server/src/simcore_service_api_server/api/routes/solvers_jobs_getters.py
@@ -41,13 +41,13 @@ from ...models.schemas.model_adapter import (
     WalletGetWithAvailableCreditsLegacy,
 )
 from ...models.schemas.solvers import SolverKeyId
+from ...models.solver_job_models_converters import create_job_from_project
 from ...services_http.director_v2 import DirectorV2Api
 from ...services_http.jobs import (
     get_custom_metadata,
     raise_if_job_not_associated_with_solver,
 )
 from ...services_http.log_streaming import LogDistributor, LogStreamer
-from ...services_http.solver_job_models_converters import create_job_from_project
 from ...services_http.solver_job_outputs import ResultsTypes, get_solver_output_results
 from ...services_http.storage import StorageApi, to_file_api_model
 from ..dependencies.application import get_reverse_url_mapper
